requirements.txt
discord==1.0.1
feedparser==6.0.2
python-dateutil==2.8.1
asyncio==3.4.3
